BaseDao: requires a type extending Identity.

+++ src/main/java/com/ibanfr/infrastructure/dao/Dao.java
@@ -1,12 +1,10 @@
 package com.ibanfr.infrastructure.dao;
 
-import com.ibanfr.domain.Identity;
-
 import java.util.Optional;
 
 public interface Dao<T>{
 
     void save(T entity);
 
-    Optional<T> findById(Integer id);
+    Optional<T> findById(Long id);
 }
